requirements.txt
pandas
numpy
mysql-connector-python
pymysql
python-dotenv
scikit-learn
#-e .
